UV/myfirstproject/pyproject.toml
[project]
name = "myfirstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
]
